requirements.txt
Flask==2.3.3
gunicorn==21.2.0
python-dotenv==1.0.0
Pillow==10.0.0
Flask-Mail==0.9.1